+++ packages/core/src/browser/dialogs.ts
@@ -385,7 +385,16 @@ export class ConfirmDialog extends AbstractDialog<boolean> {
         }
         return msg;
     }
+}
 
+export async function confirmExit(): Promise<boolean> {
+    const safeToExit = await new ConfirmDialog({
+        title: nls.localize('theia/core/quitTitle', 'Are you sure you want to quit?'),
+        msg: nls.localize('theia/core/quitMessage', 'Any unsaved changes will not be saved.'),
+        ok: Dialog.YES,
+        cancel: Dialog.NO,
+    }).open();
+    return safeToExit === true;
 }
 
 @injectable()